Morseview/requirements.txt
bcrypt==3.2.0
bidict==0.22.1
blinker==1.4
cachelib==0.9.0
cachetools==5.2.1
certifi==2022.12.7
cffi==1.14.5
charset-normalizer==3.0.1
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Email==1.4.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Session==0.4.0
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==21.2.0
idna==3.1
importlib-metadata==4.0.1
is-disposable-email==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==23.2
Pillow==8.2.0
pipdeptree==2.9.6
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.2
RPi.GPIO==0.7.0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.13
sqlparse==0.4.3
termcolor==2.1.1
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==1.0.1
WTForms==2.3.3
wtforms-validators==1.0.0
zipp==3.4.1
